MI-Practical/requirements.txt
fastapi
uvicorn
pydantic
pymongo
aiofiles
python-dotenv
ffmpeg-python